+++ src/repositories/sample.repository.ts
@@ -1,0 +1,47 @@
+import { randomUUID } from 'crypto';
+
+import type { TSample } from '@/types/sample';
+
+const samples = new Map<string, string>();
+
+export function getSamples(): TSample[] {
+  return Array.from(samples.entries()).map(([id, name]) => ({ id, name }));
+}
+
+export function createSample(name: string): TSample {
+  const id = randomUUID();
+  samples.set(id, name);
+  return { id, name };
+}
+
+export function getSampleById(id: string) {
+  const name = samples.get(id);
+  return name ? { id, name } : null;
+}
+
+export function getSampleByName(name: string) {
+  for (const [id, sampleName] of samples.entries()) {
+    if (sampleName === name) {
+      return { id, name: sampleName };
+    }
+  }
+  return null;
+}
+
+export function updateSampleById(id: string, name: string) {
+  if (!checkSampleExistsById(id)) return null;
+
+  samples.set(id, name);
+  return { id, name };
+}
+
+export function deleteSampleById(id: string) {
+  if (!checkSampleExistsById(id)) return null;
+
+  samples.delete(id);
+  return id;
+}
+
+function checkSampleExistsById(id: string): boolean {
+  return samples.has(id);
+}
